make negative results display correctly (rtl)
because calculator display is direction: rtl, negatives would end up on
the wrong side. Edit screen.textContent to put negative on the other
side, while preserving the correctly (and normally) formatted number in
the screenNumber variable

this fix consequently addresses the issue:
"delete not working on negative result, ig cuz "-" is not valid and it
checks only the first char"
because delete checks the first index to ensure it is a number, and
now the first index of screen.textContent on a negative is not the sign,
rather the sign is the last index.